Fix vault import with different RP IDs
When a vault contained multiple credential sources with different RP
IDs, only the most recently added source could successfully be used
after import.

This is because the IdentityVault Import function took the address of a
string that resided in the copy created by a range loop. Start address
of this copy is always the same for all iterations (iow, range
overwrites the previous element copy). We therefore stored the same
address &source.RelyingParty multiple times in vault.CredentialSources
and eventually virtual-fido behaves as if only the last RP ID was
imported.

The simple fix is to not use the range copies, but the range index.